// src/types/proyects.ts
export interface Project {
    image: string;
    title: string;
    description: string;
    date: string;
    type: string;
    client: string;
    duration: string;
    teamSize: string;
    stack: string[];
    architecture: string;
    integrations: string[];
    role: string;
    responsibilities: string[];
    technicalDecisions: string[];
    methodology: string;
    versionControl: string;
    testing: string;
    metrics: string[];
    problemSolved: string;
    status: string;
    challenges: string[];
    learnings: string[];
    repo?: string;
    demo?: string;
}